requirements.txt
rich==13.3.5
simple-salesforce==1.12.3
python-dotenv==1.0.0
